+++ Common/Map.h
@@ -32,7 +32,7 @@ public:
 
 	bool intersectsWall(const sf::Vector2<float> & p_position);
 
-	std::vector<sf::Vector2<int>> m_spawnPositions;
+	std::vector<sf::Vector2<float>> m_spawnPositions;
 };
 
 	sf::Packet& operator <<(sf::Packet& packet, const Type& type);
